Specify search attributes in IMAP module
Search attributes must be specified in IMAP module because it varies
according to operation being executed. For simple login we need to look
for 'mailHost' attribute only. But 'authenticate' operation needs Skeeter
to validate client's credentials before authenticating with admin account
on user's behalf. In such cases a 'password' attribute is requested.